{"schema":"decodex/commit/1","summary":"Modularize frozen prepared export store","authority":"manual"}
